- changing the name for the test file and move the files to the navbar folder
