+++ apps/modernization-ui/src/shared/header/NavBar.tsx
@@ -3,7 +3,7 @@ import { useLocation } from 'react-router-dom';
 import { Config } from '../../config';
 import { UserContext } from '../../providers/UserContext';
 import './NavBar.scss';
-import { useNavigationBarPermissions } from 'permissions';
+import { useNavigationBarPermissions } from 'shared/header/permissions';
 
 // eslint-disable-next-line no-undef
 const NBS_URL = Config.nbsUrl;
