+++ apps/modernization-ui/src/shared/header/permissions/permissions.ts
