+++ apps/modernization-ui/src/shared/header/permissions/useNavigationBarPermissions.test.tsx
